fix(showcase): repair auth 404, voice 500, and multimodal LFS deploys
Three runtime regressions from the PR #4271 shipment, propagated to all
three LangGraph showcase packages (langgraph-python, langgraph-fastapi,
langgraph-typescript).

1. Auth demo returned 404 on /info. createCopilotRuntimeHandler (V2)
   URL-routes on /info, /agent/:id/run, etc. It needs a Next.js catchall
   route so sub-paths reach it. PR #4271 mounted it at the fixed-path
   route.ts, which meant every sub-path 404'd. Moved to [[...slug]]/route.ts.

2. Voice demo returned 500 on POST /transcribe. V1 CopilotRuntime's
   transcriptionService is a getter-only delegate; the PR-#4271 pattern
   v2Instance.transcriptionService = ... crashes. Rewrote the voice
   route to use V2 CopilotRuntime directly from @copilotkit/runtime/v2
   and pass transcriptionService in the constructor. V2 handler also
   URL-routes so also moved to [[...slug]]/route.ts.

3. Multimodal LFS-pointer deploy failure. /demos/multimodal sample
   buttons threw a pointer error because the deployed dev server served
   the 130-byte LFS stub when git lfs pull wasn't run in the Docker
   build. Resolved LFS and added a package-scoped .gitattributes that
   exempts the three demo samples (sample.png, sample.pdf, sample.wav)
   from LFS tracking via -filter -diff -merge binary. Committed as real
   binaries so deploy environments see the actual asset regardless of
   LFS configuration. The client-side pointer guard stays in place as a
   defensive net.

All three fixes applied identically across the three showcase packages.
