+++ showcase/packages/langgraph-fastapi/src/app/api/copilotkit-auth/[[...slug]]/route.ts
