+++ showcase/packages/langgraph-python/src/app/api/copilotkit-auth/[[...slug]]/route.ts
