+++ showcase/packages/langgraph-typescript/src/app/api/copilotkit-auth/[[...slug]]/route.ts
